+++ .packages
@@ -1,4 +1,4 @@
-# Generated by pub on 2021-02-18 10:49:05.006492.
+# Generated by pub on 2021-02-18 12:47:29.081074.
 async:file:///C:/src/flutter/flutter/.pub-cache/hosted/pub.dartlang.org/async-2.5.0-nullsafety.1/lib/
 boolean_selector:file:///C:/src/flutter/flutter/.pub-cache/hosted/pub.dartlang.org/boolean_selector-2.1.0-nullsafety.1/lib/
 characters:file:///C:/src/flutter/flutter/.pub-cache/hosted/pub.dartlang.org/characters-1.1.0-nullsafety.3/lib/
