--- lib/constants.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-const kLabelTextStyleBlue = TextStyle(
-    fontSize: 18.0,
-    color: Color(0xFF0B0F82),
-    fontWeight: FontWeight.bold,
-    fontFamily: 'Poppins');
-
-const kLabelTextStyleYellow = TextStyle(
-  fontSize: 18.0,
-  color: Color(0xFFF4D50A),
-);
