+++ lib/screens/login.dart
@@ -1,7 +1,6 @@
 import 'package:fast_shop/components/constants.dart';
 import 'package:fast_shop/components/inputfield.dart';
 import 'package:fast_shop/components/square_button.dart';
-import 'package:fast_shop/constants.dart';
 import 'package:flutter/material.dart';
 
 class LoginPage extends StatefulWidget {
